testingphase1.0

+++ src/app/modules/manager/manager.component.ts
@@ -5,19 +5,21 @@ import { RouterLink } from '@angular/router';
 import Toppings from 'app/interface/topping';
 import {MatCardModule} from '@angular/material/card';
 import { NgFor, JsonPipe, NgIf} from '@angular/common';
-import { FormBuilder, FormsModule, ReactiveFormsModule, FormGroup, FormControl } from '@angular/forms';
+import { FormBuilder, FormsModule, ReactiveFormsModule, FormGroup, NgForm } from '@angular/forms';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {MatInputModule} from '@angular/material/input';
 import { ToppingService } from 'app/topping.service';
+import { MatDialog, MatDialogModule} from '@angular/material/dialog';
+import { DeleteConfirmDialog } from '../dialog/delete-confirm-dialog';
 @Component({
     selector     : 'manager-component',
     standalone   : true,
     templateUrl  : './manager.component.html',
     encapsulation: ViewEncapsulation.None,    
     imports      : [MatButtonModule, RouterLink, MatIconModule, MatCardModule, NgFor, NgIf,
-        FormsModule, MatCheckboxModule, MatSidenavModule, MatFormFieldModule, MatInputModule,
+        FormsModule, MatCheckboxModule, MatSidenavModule, MatFormFieldModule, MatInputModule, MatDialogModule,
          ReactiveFormsModule, JsonPipe],
 })
 export class managerComponent implements  OnInit 
@@ -40,7 +42,9 @@ export class managerComponent implements  OnInit
      * Constructor
      */
 
-    constructor(private _formBuilder: FormBuilder, private _toppingService: ToppingService)
+    constructor(private _formBuilder: FormBuilder, 
+        private _toppingService: ToppingService, 
+        public deleteConfirmDialog: MatDialog)
     {}
 
 
@@ -54,6 +58,28 @@ export class managerComponent implements  OnInit
         )   
     
     }
+
+    submitForm(form: NgForm) {
+
+        console.log(form);
+        console.log(this.topping);
+      
+        if (form.valid) {
+          // Handle form submission
+          console.log('Form submitted successfully!');
+
+          //check the pizza name can't repeat
+          
+
+          if(this.topping.id){
+            this.UpdateTopping();
+          }
+          else{
+            this.CreateTopping();
+          }
+
+        }
+      }
 
     CreateTopping(){
 
@@ -103,13 +129,26 @@ export class managerComponent implements  OnInit
 
     DeleteTopping(_topping:Toppings){
 
-        this._toppingService.deleteTopping(_topping).subscribe(
-            response=>{
-                //mock
-                 let index = this.toppingList.findIndex(p=>p.id===_topping.id);
-                 this.toppingList.splice(index,1);
+        const dialogRef = this.deleteConfirmDialog.open(DeleteConfirmDialog, {
+            width: '250px',
+          });
+      
+          dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+
+              console.log(result);
+              // Call your delete function here
+              this._toppingService.deleteTopping(_topping).subscribe(
+                response=>{
+                    //mock
+                     let index = this.toppingList.findIndex(p=>p.id===_topping.id);
+                     this.toppingList.splice(index,1);
+                }
+            );
+              console.log('Delete confirmed');
             }
-        );
+          });
+
     }
 
     Reset(){
